Build: fine grained build targets for hesai driver
