+++ modules/drivers/hesai/hesai_component.h
@@ -14,51 +14,24 @@
  * limitations under the License.
  *****************************************************************************/
 
-#ifndef LIDAR_HESAI_HESAI_ALL_COMPONENT_H_
-#define LIDAR_HESAI_HESAI_ALL_COMPONENT_H_
+#pragma once
 
-#include <list>
 #include <memory>
 #include <string>
-#include <thread>
 
-#include "cyber/cyber.h"
-#include "modules/drivers/hesai/const_var.h"
-#include "modules/drivers/hesai/driver.h"
-#include "modules/drivers/hesai/parser.h"
 #include "modules/drivers/hesai/proto/config.pb.h"
-#include "modules/drivers/hesai/type_defs.h"
+#include "cyber/cyber.h"
+#include "modules/drivers/hesai/driver.h"
+#include "modules/drivers/hesai/parser_factory.h"
 
 namespace apollo {
 namespace drivers {
 namespace hesai {
 
-using apollo::cyber::Component;
-
-class HesaiComponent : public Component<> {
+class HesaiComponent : public ::apollo::cyber::Component<> {
  public:
   ~HesaiComponent() {}
-  bool Init() override {
-    if (!GetProtoConfig(&conf_)) {
-      AERROR << "load config error, file:" << config_file_path_;
-      return false;
-    }
-
-    AINFO << "conf:" << conf_.DebugString();
-    Parser* parser = ParserFactory::CreateParser(node_, conf_);
-    if (parser == nullptr) {
-      AERROR << "create parser error";
-      return false;
-    }
-    parser_.reset(parser);
-    driver_.reset(new HesaiDriver(node_, conf_, parser_));
-
-    if (!driver_->Init()) {
-      AERROR << "driver init error";
-      return false;
-    }
-    return true;
-  }
+  bool Init() override;
 
  private:
   std::shared_ptr<HesaiDriver> driver_;
@@ -71,5 +44,3 @@ CYBER_REGISTER_COMPONENT(HesaiComponent)
 }  // namespace hesai
 }  // namespace drivers
 }  // namespace apollo
-
-#endif
